Renamed confusing menu package to main
