+++ LudoCompose/app/src/main/java/hu/bme/aut/android/ludocompose/navigation/main/Screen.kt
@@ -14,22 +14,18 @@
  * limitations under the License.
  */
 
-package hu.bme.aut.android.ludocompose.navigation.menu
+package hu.bme.aut.android.ludocompose.navigation.main
 
 import androidx.annotation.StringRes
 import hu.bme.aut.android.ludocompose.R
 import hu.bme.aut.android.ludocompose.navigation.common.Screen
 
-sealed class MainScreen(
-    override val route: String,
-    @StringRes override val title: Int
-) : Screen
-data object MenuScreen : MainScreen("menu", R.string.menu_main) {
+data object MenuScreen : Screen("main_menu", R.string.menu_main) {
     override val enableNavigationBack: Boolean get() = false
 }
 
-data object LocalMenuScreen : MainScreen("local_menu", R.string.menu_local)
+data object LocalMenuScreen : Screen("local_menu", R.string.menu_local)
 
-data object OnlineMenuScreen : MainScreen("online_menu", R.string.menu_online)
+data object OnlineMenuScreen : Screen("online_menu", R.string.menu_online)
 
-data object AboutScreen : MainScreen("about", R.string.menu_about)
+data object AboutScreen : Screen("about", R.string.menu_about)
